project/pyproject.toml
[tool.black]
exclude = '''
(
  /(
      \.mypy_cache
    | \.pytest_cache
    | migrations
  )/
)
'''
line-length = 119

[tool.flake8]
exclude = [".mypy_cache", ".pytest_cache", "migrations"]
max-line-length = 119

[tool.isort]
profile = "black"
skip_glob = [".mypy_cache", ".pytest_cache", "migrations"]

[tool.poetry]
authors = ["Evgeni Sautin <evgeni.sautin@gmail.com>"]
description = "Test-driven development with FastAPI and Docker"
name = "project"
version = "0.1.0"

[tool.poetry.dependencies]
aerich = "^0.5.3"
asyncpg = "^0.22.0"
fastapi = "^0.63.0"
gunicorn = "^20.1.0"
newspaper3k = "^0.2.8"
python = "^3.9"
requests = "^2.25.1"
tortoise-orm = "^0.17.2"
uvicorn = "^0.13.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake9 = "^3.8.3"
isort = "^5.8.0"
mypy = "^0.812"
pylint = "^2.7.4"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
pytest-xdist = "^2.2.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
